Phase 8f: bench-wasm preset — Emscripten + Highway SIMD128
Lets us measure m_bench under Node.js against a native AVX2 build.
The only numkit code change is an explicit size_t cast in MStdMatrix
reshape — upstream clang (emscripten) flags int→size_t narrowing as
error where MSVC allows it silently.

Build/run:
  cmake --preset=bench-wasm
  cmake --build --preset=bench-wasm
  node build-bench-wasm/benchmarks/m_bench.js [--benchmark_filter=...]

Infrastructure changes:
  - CMakePresets.json: new bench-wasm inheriting from browser with
    -msimd128 -mrelaxed-simd, NUMKIT_M_BUILD_BENCHMARKS=ON, SIMD=ON.
    Tests / REPL / install disabled to keep the build focused.
  - CMakeLists.txt: dropped NOT-EMSCRIPTEN guard on the benchmarks
    subdirectory and on BENCHMARK_ENABLE_WERROR. Also append
    -Wno-c2y-extensions to the benchmark / benchmark_main targets;
    v1.8.3 hits a C2y __COUNTER__ extension under clang 19 which
    -pedantic-errors turns into a hard error.
  - benchmarks/CMakeLists.txt: Emscripten branch sets m_bench's
    suffix to .js, enables -fexceptions, and picks stable runtime
    flags (ALLOW_MEMORY_GROWTH, INITIAL_MEMORY=128MiB, NODERAWFS,
    EXIT_RUNTIME, STACK_SIZE=4MiB, DISABLE_EXCEPTION_CATCHING=0).

Tests: 3634/3634 green on desktop-fast (native unaffected). WASM
benchmark runs clean under node 22. Observed numbers are written
up in project_hw_toolchain memory.

+++ libs/builtin/src/MStdMatrix.cpp
@@ -384,7 +384,7 @@ void reshape_reg(Span<const MValue> args, size_t /*nargout*/, Span<MValue> outs,
         // Scalar-args form: reshape(A, m, n) or (A, m, n, p). One [] allowed
         // for dimension inference from x.numel().
         const size_t dimCount = args.size() - 1;
-        size_t dims[3] = {1, 1, (dimCount >= 3) ? 1 : 0};
+        size_t dims[3] = {1, 1, static_cast<size_t>((dimCount >= 3) ? 1 : 0)};
         int inferPos = -1;
         size_t knownProd = 1;
         for (size_t i = 0; i < dimCount && i < 3; ++i) {
